Initial set of messages for changes of the phases
First string contains message that the spy should recieve instead of the chosen question.
Second string contains the message diplomats should recieve at the end of teh carousel round.
third string contains the message the spy should recieve at the end of the carousel round.
Fourth string contains the message for start of the tribunal. Everyone should recieve this one.
Fifth string contains the message for end of tribunal. Everyone should recieve this one.
Sixth string contains the message for the results of the tribunal. Everyone should recieve this but only if the spy survives.